#195 Refactor: split input to a input search view and input search controller
